feat(FIN-048): refactor imports to use new auth directory structure and update loading state checks in guards
