+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { NotAuthGuard } from '@frontend/entities/profile/not-auth-guard';
+import { NotAuthGuard } from '@frontend/entities/auth/not-auth-guard';
 import { WelcomeScreen } from '@frontend/features/welcome/welcome-screen';
 
 export default function MainPage() {
