+++ src/client/entities/auth/auth-guard.tsx
@@ -5,7 +5,7 @@ import { useEffect } from 'react';
 import { useUserInformation } from '@frontend/shared/services/user-information/use-user-information.store';
 import { useShallow } from 'zustand/react/shallow';
 import { PromiseState } from '@frontend/shared/enums/promise-state.enum';
-import type { AuthGuardProps } from '@frontend/entities/profile/props/auth-guard.props';
+import type { AuthGuardProps } from '@frontend/entities/auth/props/auth-guard.props';
 
 export function AuthGuard({ children, routePath = '/login' }: AuthGuardProps) {
   const { userInfoState, userInformation } = useUserInformation(
@@ -16,7 +16,7 @@ export function AuthGuard({ children, routePath = '/login' }: AuthGuardProps) {
   const pathName = usePathname();
 
   useEffect(() => {
-    if (userInfoState !== PromiseState.Success) {
+    if (userInfoState === PromiseState.Loading) {
       return;
     }
     if (!userInformation) {
