+++ src/client/entities/auth/authorized-user.hook.ts
