+++ src/client/entities/auth/props/auth-guard.props.ts
