+++ src/client/entities/auth/role-guard.tsx
@@ -5,7 +5,7 @@ import { useEffect } from 'react';
 import { useUserInformation } from '@frontend/shared/services/user-information/use-user-information.store';
 import { useShallow } from 'zustand/react/shallow';
 import { PromiseState } from '@frontend/shared/enums/promise-state.enum';
-import type { AuthGuardProps } from '@frontend/entities/profile/props/auth-guard.props';
+import type { AuthGuardProps } from '@frontend/entities/auth/props/auth-guard.props';
 import { RoleEnum } from '@common/domains/user/enums/role.enum';
 import { FinLoader } from '@frontend/shared/components/loader/fin-loader';
 import { FinErrorWidget } from '@frontend/components/error/fin-error-widget';
