+++ src/client/features/tracking-operation/tracking-operation-screen.tsx
@@ -22,7 +22,7 @@ import {
 import { useGetBasicTrackingInformation } from './tracking-operation-filters/tracking-operation-hooks/get-tracking-op-information.hook';
 import { useTrackingOperationFilters } from '@frontend/features/tracking-operation/tracking-operation-filters/tracking-operation-hooks/tracking-operation-filters.hook';
 import { TrackingOperationQueryKey } from '@frontend/entities/tracking-operations/tracking-operation-query-key.constant';
-import { useAuthorizedUser } from '@frontend/entities/profile/authorized-user.hook';
+import { useAuthorizedUser } from '@frontend/entities/auth/authorized-user.hook';
 import { formatDate } from '@frontend/shared/utils/format-date.util';
 import { DateFormatType } from '@frontend/shared/enums/date-type.enum';
 import { FinListPageWrapper } from '@frontend/components/wrappers/fin-list-page-wrapper';
