+++ src/client/widgets/admin-sidebar/admin-sidebar-content.tsx
@@ -10,7 +10,7 @@ import { UiSidebarMenuButton } from '@frontend/ui/ui-sidebar/ui-sidebar-menu-but
 import { UiSidebarMenuItem } from '@frontend/ui/ui-sidebar/ui-sidebar-menu-item';
 import { UiSidebarSeparator } from '@frontend/ui/ui-sidebar/ui-sidebar-separator';
 import { UiSvgIcon } from '@frontend/ui/ui-svg-icon/ui-svg-icon';
-import { useAuthorizedUser } from '@frontend/entities/profile/authorized-user.hook';
+import { useAuthorizedUser } from '@frontend/entities/auth/authorized-user.hook';
 import { adminNavRoutesWindow } from '@frontend/widgets/shared/admin-routes.constant';
 import React from 'react';
