--- src/common/domains/auth/models/login-result.model.ts
@@ -1,3 +1,0 @@
-export interface LoginResultModel {
-  success: boolean;
-}
