--- src/common/domains/auth/models/login.response.ts
@@ -1,3 +1,0 @@
-export interface LoginResponse {
-  token: string;
-}
